add example friends config with docs and some script support
Mixed feelings about this trick for installing the example
configuration. The simpler instructions would be to just have admins
edit and install the config files themselves instead of creating this
somewhat elaborate machine to "simplify" the installation of config
files.

I think this much is repeatable. But unsure if it's what we really
want in the long run.